+++ src/codplayer/player.py
@@ -216,6 +216,8 @@ class Player(object):
     #
 
     def cmd_disc(self, args):
+        source_disc_id = None
+
         if args:
             # Play disc in database by its ID
             did = args[0]
@@ -243,6 +245,14 @@ class Player(object):
             except rip.RipError, e:
                 raise CommandError('rip failed: {}'.format(e))
 
+            # Only follow links for physical discs.  When the user
+            # starts a disc by ID we assume they really want to listen
+            # to that one.
+            disc, source_disc_id = self.resolve_alias_links(disc)
+
+        # Stash the source disc into the resolved one.  Slightly ugly
+        # messing with model.DbDisc like this, but it's simple.
+        disc.source_disc_id = source_disc_id
         return self.play_disc(disc)
 
 
@@ -324,6 +334,49 @@ class Player(object):
 
     def cmd_version(self, args):
         return full_version()
+
+
+    def resolve_alias_links(self, disc):
+        """Follow any disc alias links, returning the disc that should really
+        be played.
+
+        Any errors are handled by just returning whatever disc has
+        been reached in the chain, as that is probably good enough to
+        play.
+        """
+
+        # Keep track of the disc that started the link
+        source_disc = disc
+        source_disc_id = None
+
+        # Avoid getting stuck in circles
+        visited = set()
+        visited.add(disc.disc_id)
+
+        while disc.link_type == 'alias':
+            if disc.linked_disc_id:
+                linked = self.db.get_disc_by_disc_id(disc.linked_disc_id)
+            else:
+                linked = None
+
+            if not linked:
+                self.log('error: missing alias link from {} to {}',
+                         disc, disc.linked_disc_id)
+                break
+
+            if linked.disc_id in visited:
+                self.log('error: alias link circle from {} to {}',
+                         disc, linked)
+                break
+
+            self.debug('following alias link from {} to {}',
+                       disc, linked)
+
+            visited.add(linked.disc_id)
+            disc = linked
+            source_disc_id = source_disc.disc_id
+
+        return disc, source_disc_id
 
 
     def play_disc(self, disc, track_number = 0):
@@ -703,6 +756,7 @@ class Transport(object):
     def set_state_working(self):
         self.state.state = State.WORKING
         self.state.disc_id = self.source.disc.disc_id
+        self.state.source_disc_id = self.source.disc.source_disc_id
         self.state.track = self.start_track + 1
         self.state.no_tracks = len(self.source.disc.tracks)
         self.state.index = 0